requirements.txt
langchain
langchain-community
pypdf
python-docx
faiss-cpu
sentence-transformers
transformers
huggingface_hub
aisuite[all]
gradio
python-dotenv
